# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  /** @minItems 1 */
  categories: string[];
  tags?: string[];
  /** @minimum 0.01 */
  basePrice: number;
  /**
   * @minimum 1
   * @maximum 180
   */
  deliveryDays: number;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface RefreshTokenBody {
  refreshToken: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
  user: AuthUser;
}

export interface RefreshTokenResponse {
  accessToken: string;
  expiresIn: number;
}

export interface Artist {
  id: string;
  name: string;
  email: string;
  categories: string[];
  tags: string[];
  basePrice: number;
  deliveryDays: number;
  availability: boolean;
  rating: number;
  totalReviews: number;
  createdAt: string;
}

export interface ArtistListResponse {
  artists: Artist[];
  total: number;
}

export interface UpdateArtistBody {
  name?: string;
  categories?: string[];
  tags?: string[];
  basePrice?: number;
  deliveryDays?: number;
}

export interface ToggleAvailabilityBody {
  availability: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  PROPOSED: "PROPOSED",
  PAYMENT_PENDING: "PAYMENT_PENDING",
  PAID: "PAID",
  IN_PROGRESS: "IN_PROGRESS",
  DELIVERED: "DELIVERED",
  CANCELLED: "CANCELLED",
} as const;

export interface Order {
  id: string;
  clientId: string;
  artistId: string;
  title: string;
  description?: string;
  occasion?: string;
  names?: string[];
  referenceLinks?: string[];
  deadline: string;
  additionalInstructions?: string;
  basePrice: number;
  status: OrderStatus;
  clientName?: string;
  clientEmail?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface OrderListResponse {
  orders: Order[];
  total: number;
}

export interface CreateOrderBody {
  artistId: string;
  clientName: string;
  clientEmail: string;
  title: string;
  description?: string;
  occasion?: string;
  names?: string[];
  referenceLinks?: string[];
  deadline: string;
  additionalInstructions?: string;
  basePrice?: number;
}

export interface UpdateOrderStatusBody {
  status: OrderStatus;
  deliveryVideoUrl?: string;
}

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus];

export const PaymentStatus = {
  PENDING: "PENDING",
  CONFIRMED: "CONFIRMED",
  REFUNDED: "REFUNDED",
  FAILED: "FAILED",
} as const;

export type PaymentProvider =
  (typeof PaymentProvider)[keyof typeof PaymentProvider];

export const PaymentProvider = {
  stripe: "stripe",
  asaas: "asaas",
} as const;

export interface Payment {
  id: string;
  orderId: string;
  amount: number;
  currency: string;
  status: PaymentStatus;
  provider: PaymentProvider;
  transactionId?: string;
  createdAt: string;
}

export type CreateCheckoutBodyProvider =
  (typeof CreateCheckoutBodyProvider)[keyof typeof CreateCheckoutBodyProvider];

export const CreateCheckoutBodyProvider = {
  stripe: "stripe",
  asaas: "asaas",
} as const;

export interface CreateCheckoutBody {
  orderId: string;
  provider?: CreateCheckoutBodyProvider;
}

export interface CheckoutResponse {
  checkoutUrl: string;
  sessionId: string;
  expiresAt?: string;
}

export type ConfirmPaymentBodyProvider =
  (typeof ConfirmPaymentBodyProvider)[keyof typeof ConfirmPaymentBodyProvider];

export const ConfirmPaymentBodyProvider = {
  stripe: "stripe",
  asaas: "asaas",
} as const;

export interface ConfirmPaymentBody {
  sessionId: string;
  provider: ConfirmPaymentBodyProvider;
}

export interface Media {
  id: string;
  artistId: string;
  fileName: string;
  fileSize: number;
  fileUrl: string;
  mimeType: string;
  uploadedAt: string;
}

export interface MediaListResponse {
  media: Media[];
  total: number;
}

export interface Review {
  id: string;
  orderId: string;
  artistId: string;
  clientId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
  createdAt: string;
}

export interface ReviewListResponse {
  reviews: Review[];
  total: number;
  averageRating: number;
}

export interface CreateReviewBody {
  orderId: string;
  clientId: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /** @maxLength 500 */
  comment?: string;
}

export interface OrderStatusCount {
  status: OrderStatus;
  count: number;
}

export interface DashboardStats {
  totalEarnings: number;
  pendingEarnings: number;
  totalOrders: number;
  ordersByStatus: OrderStatusCount[];
  rating: number;
  totalReviews: number;
  availability: boolean;
  completionRate: number;
}

export interface RecentOrder {
  id: string;
  title: string;
  clientName: string;
  clientEmail?: string;
  status: OrderStatus;
  basePrice: number;
  deadline: string;
  daysRemaining: number;
  createdAt: string;
}

export interface RecentOrdersResponse {
  orders: RecentOrder[];
}

export interface MonthlyEarning {
  month: string;
  year: number;
  earnings: number;
  orderCount: number;
}

export interface EarningsResponse {
  monthly: MonthlyEarning[];
  totalEarnings: number;
  averageMonthly: number;
}

export type ListArtistsParams = {
  category?: string;
  available?: boolean;
  minPrice?: number;
  maxPrice?: number;
  limit?: number;
  offset?: number;
};

export type ListOrdersParams = {
  status?: ListOrdersStatus;
  limit?: number;
  offset?: number;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  PROPOSED: "PROPOSED",
  PAYMENT_PENDING: "PAYMENT_PENDING",
  PAID: "PAID",
  IN_PROGRESS: "IN_PROGRESS",
  DELIVERED: "DELIVERED",
  CANCELLED: "CANCELLED",
} as const;

export type UploadMediaBody = {
  file: Blob;
};

export type GetArtistReviewsParams = {
  limit?: number;
  offset?: number;
};

export type GetRecentOrdersParams = {
  limit?: number;
};

export type GetEarningsParams = {
  months?: number;
};