# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Artists Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Artist registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login artist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenBody"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout artist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ARTISTS
  /artists/me:
    get:
      operationId: getMe
      tags: [artists]
      summary: Get current artist profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateMe
      tags: [artists]
      summary: Update current artist profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists/me/availability:
    patch:
      operationId: toggleAvailability
      tags: [artists]
      summary: Toggle artist availability
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleAvailabilityBody"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: List all artists (public)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of artists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistListResponse"
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get artist by ID (public)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ORDERS
...
  /media:
    get:
      operationId: listMedia
      tags: [media]
      summary: List media for current artist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of media
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"

    post:
      operationId: uploadMedia
      tags: [media]
      summary: Upload a video file (max 25MB)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Media uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "400":
          description: File too large or invalid type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /media/{id}:
    delete:
      operationId: deleteMedia
      tags: [media]
      summary: Delete media file
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # REVIEWS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        categories:
          type: array
          items:
            type: string
          minItems: 1
        tags:
          type: array
          items:
            type: string
        basePrice:
          type: number
          minimum: 0.01
        deliveryDays:
          type: integer
          minimum: 1
          maximum: 180
      required:
        - name
        - email
        - password
        - categories
        - basePrice
        - deliveryDays
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RefreshTokenBody:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
      required:
        - id
        - email
        - name
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
    RefreshTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
      required:
        - accessToken
        - expiresIn

    # ARTIST SCHEMAS
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        basePrice:
          type: number
        deliveryDays:
          type: integer
        availability:
          type: boolean
        rating:
          type: number
        totalReviews:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - categories
        - tags
        - basePrice
        - deliveryDays
        - availability
        - rating
        - totalReviews
        - createdAt
    ArtistListResponse:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/Artist"
        total:
          type: integer
      required:
        - artists
        - total
    UpdateArtistBody:
      type: object
      properties:
        name:
          type: string
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        basePrice:
          type: number
        deliveryDays:
          type: integer
    ToggleAvailabilityBody:
      type: object
      properties:
        availability:
          type: boolean
      required:
        - availability

    # ORDER SCHEMAS
...
    Media:
      type: object
      properties:
        id:
          type: string
        artistId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        fileUrl:
          type: string
        mimeType:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - artistId
        - fileName
        - fileSize
        - fileUrl
        - mimeType
        - uploadedAt
    MediaListResponse:
      type: object
      properties:
        media:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        total:
          type: integer
      required:
        - media
        - total

    # REVIEW SCHEMAS
...