# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Media {
  id: string;
  artistId: string;
  fileName: string;
  fileSize: number;
  fileUrl: string;
  mimeType: string;
  uploadedAt: Date;
}