# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Media } from "./media";

export interface MediaListResponse {
  media: Media[];
  total: number;
}