# lib/api-zod/src/generated/types/refreshTokenResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RefreshTokenResponse {
  accessToken: string;
  expiresIn: number;
}